requirements.txt
accelerate==1.11.0
audioread==3.1.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
decorator==5.2.1
einops==0.8.1
ffmpeg-python==0.2.0
filelock==3.20.0
fsspec==2025.10.0
future==1.0.0
huggingface-hub==0.36.0
idna==3.11
inflect==7.5.0
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.2
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.45.1
loguru==0.7.3
MarkupSafe==3.0.3
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
networkx==3.5
nltk==3.9.2
numba==0.62.1
numpy==2.3.4
packaging==25.0
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
pooch==1.8.2
progressbar==2.5
psutil==7.1.3
pycparser==2.23
Pygments==2.19.2
pyphen==0.17.2
pytest==9.0.1
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
rotary-embedding-torch==0.8.9
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==5.1.2
sentencepiece==0.2.1
soundfile==0.13.1
soxr==1.0.0
sympy==1.14.0
textstat==0.7.11
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.1
torchaudio==2.9.1+cpu
tortoise-tts==3.0.0
tqdm==4.67.1
transformers==4.57.1
typeguard==4.4.4
typing_extensions==4.15.0
Unidecode==1.4.0
urllib3==2.5.0
win32_setctime==1.2.0
